Guard against thundering herds towards DNS
If we hit a nxdomain situation, the system would crash and attempt a
reconnect ASAP. This creates yet another request toward the DNS
server, another crash, yet another request and so on. This allows the
system to quickly overwhelm a DNS server with traffic.

The fix is to detect this situation and then back off for a minute.
nxdomain updates tend to take some time anyway, so it isn't that
important to try again immediately. We can await the state of the DNS
to change. In the meantime, the systems buffering should handle the
case and do things the right way.